+++ lib/src/apps/downloadable_apps/downloadable_apps.dart
@@ -1,6 +1,7 @@
 export 'facebook.dart';
 export 'instagram.dart';
 export 'linkedin.dart';
+export 'pinterest.dart';
 export 'telegram.dart';
 export 'twitter.dart';
 export 'whatsapp.dart';